\name{add_button}

\alias{add_button}
\alias{add_button2}
\alias{add_button3}
\alias{add_close}
\alias{add_close2}
\alias{add_close3}
\alias{draw_flags}
\alias{filter_region}
\alias{map_ata}
\alias{map_country}
\alias{map_flag}
\alias{map_focus}
\alias{map_guess}
\alias{map_world}
\alias{mask_flag}
\alias{new_dev}
\alias{onclic}
\alias{onclic2}
\alias{onclic3}

\title{geographr internal functions}

\description{geographr internal functions}

\details{These functions are low level functions. Not for a direct use.}

\author{Nicolas Casajus}
