% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_ccg_21}
\alias{boundaries_ccg_21}
\title{Clinical Commissioning Groups (April 2021)}
\format{
A data frame of class "sf" with 106 rows and 3 variables:
\describe{
\item{ccg_21_name}{CCG name}
\item{ccg_21_code}{CCG code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_ccg_21
}
\description{
A data set containing 2021 Clinical Commissioning Group (CCG) geometries and
their associated names and codes. Boundaries are ultra generalised (500m).
}
\keyword{datasets}
