% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_countries20}
\alias{boundaries_countries20}
\title{Countries (December 2020)}
\format{
A data frame of class "sf" with 4 rows and 3 variables:
\describe{
\item{country20_name}{country name}
\item{country20_code}{Country code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_countries20
}
\description{
A data set containing 2020 Country geometries and
their associated names and codes. Boundaries are Ultra Generalised (500m) -
clipped to the coastline (Mean High Water mark).
}
\keyword{datasets}
