% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_dz_11}
\alias{boundaries_dz_11}
\title{Data Zones (2011)}
\format{
A data frame of class "sf" with 6,976 rows and 3 variables:
\describe{
\item{dz_11_name}{Data Zone name}
\item{dz_11_code}{Data Zone code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://spatialdata.gov.scot/}
}
\usage{
boundaries_dz_11
}
\description{
A data set containing 2011 Scottish Data Zone geometries and their associated
names and codes.
}
\keyword{datasets}
