% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_epg04}
\alias{boundaries_epg04}
\title{Emergency Preparedness Groups, Northern Ireland (2004)}
\format{
A data frame of class "sf" with 3 rows and 2 variables:
\describe{
\item{epg04_name}{Emergency Preparedness Greoup name}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk//}
}
\usage{
boundaries_epg04
}
\description{
A data set containing 2004 NI Emergency Preparedness Group geometries and their associated
names. Boundaries created by dissolving LGD geometries, from ONS.
}
\keyword{datasets}
