% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_hb19}
\alias{boundaries_hb19}
\title{Health Boards (2019)}
\format{
A data frame of class "sf" with 14 rows and 3 variables:
\describe{
\item{hb19_name}{Health Board name}
\item{hb19_code}{Health Board code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://spatialdata.gov.scot/}
}
\usage{
boundaries_hb19
}
\description{
A data set containing 2019 NHS Health Board geometries and their associated
names and codes.
}
\keyword{datasets}
