% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_icb22}
\alias{boundaries_icb22}
\title{Integrated Care Boards (2022)}
\format{
A data frame of class "sf" with 42 rows and 3 variables:
\describe{
\item{icb22_name}{Integrated Care Board name}
\item{icb22_code}{Integrated Care Board code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_icb22
}
\description{
A data set containing July 2022 NHS Integrated Care Board geometries and
their associated names and codes.
}
\keyword{datasets}
