% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_iz_11}
\alias{boundaries_iz_11}
\title{Intermediate Zones (2011)}
\format{
A data frame of class "sf" with 1,279 rows and 3 variables:
\describe{
\item{iz_11_name}{Intermediate Zone name}
\item{iz_11_code}{Intermediate Zone code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://spatialdata.gov.scot/}
}
\usage{
boundaries_iz_11
}
\description{
A data set containing 2011 Scottish Intermediate Zone geometries and their
associated names and codes.
}
\keyword{datasets}
