% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_lhb22}
\alias{boundaries_lhb22}
\title{Wales Local Health Boards (2022)}
\format{
A data frame of class "sf" with 7 rows and 3 variables:
\describe{
\item{lhb22_name}{Local Health Board name}
\item{lhb22_code}{Local Health Board code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_lhb22
}
\description{
A data set containing 2022 Local Health Board geometries for Wales and their
associated names and codes.
}
\keyword{datasets}
