% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_lhb_20}
\alias{boundaries_lhb_20}
\title{Wales Local Health Boards (2020)}
\format{
A data frame of class "sf" with 7 rows and 3 variables:
\describe{
\item{lhb_20_name}{Local Health Board name}
\item{lhb_20_code}{Local Health Board code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_lhb_20
}
\description{
A data set containing 2020 Local Health Board geometries for Wales and their
associated names and codes.
}
\keyword{datasets}
