% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_lrf20}
\alias{boundaries_lrf20}
\title{England and Wales Local Resilience Forums (2020)}
\format{
A data frame of class "sf" with 42 rows and 3 variables:
\describe{
\item{lrf20_name}{Local Resilience Forum name}
\item{lrf20_code}{Local Resilience Forum code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_lrf20
}
\description{
A data set containing 2020 Local Resilience Forum geometries for England and Wales and their
associated names and codes.
}
\keyword{datasets}
