% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_lrp19}
\alias{boundaries_lrp19}
\title{Scotland Local Resilience Partnerships (2019)}
\format{
A data frame of class "sf" with 12 rows and 3 variables:
\describe{
\item{lrp19_name}{Local Resilience Partnership name}
\item{lrp19_code}{Local Resilience Partnership code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://data.gov.uk/}
}
\usage{
boundaries_lrp19
}
\description{
A data set containing 2019 Local Resilience Partnership geometries for Scotland and their
associated names and codes.
}
\keyword{datasets}
