% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_lsoa_11}
\alias{boundaries_lsoa_11}
\title{Lower Layer Super Output Areas (2011)}
\format{
A data frame of class "sf" with 34,076 rows and 3 variables:
\describe{
\item{lsoa_11_name}{LSOA name}
\item{lsoa_11_code}{LSOA code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_lsoa_11
}
\description{
A data set containing 2011 Lower Super Output Area (LSOA) geometries and
their associated names and codes. Full clipped boundaries were simplified
using the Visvalingam algorithm set to a weighting of 0.7 with a retaining
proportion of 0.05 points.
}
\keyword{datasets}
