% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_ltla19}
\alias{boundaries_ltla19}
\title{Local Authority Districts (December 2019)}
\format{
A data frame of class "sf" with 382 rows and 3 variables:
\describe{
\item{ltla19_name}{LAD name}
\item{ltla19_code}{LAD code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_ltla19
}
\description{
A data set containing 2019 Local Authority Districts (LAD) geometries and
their associated names and codes. Boundaries are Ultra Generalised (500m) -
clipped to the coastline (Mean High Water mark).
}
\keyword{datasets}
