% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_ltla20}
\alias{boundaries_ltla20}
\title{Local Authority Districts (December 2020)}
\format{
A data frame of class "sf" with 379 rows and 3 variables:
\describe{
\item{ltla20_name}{LAD name}
\item{ltla20_code}{LAD code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_ltla20
}
\description{
A data set containing 2020 Local Authority Districts (LAD) geometries and
their associated names and codes. Boundaries are Ultra Generalised (500m) -
clipped to the coastline (Mean High Water mark).
}
\keyword{datasets}
