% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_ltla21}
\alias{boundaries_ltla21}
\title{Local Authority Districts (December 2021)}
\format{
A data frame of class "sf" with 374 rows and 3 variables:
\describe{
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_ltla21
}
\description{
A data set containing 2021 Local Authority Districts (LAD) geometries and
their associated names and codes. Boundaries are Ultra Generalised (500m) -
clipped to the coastline (Mean High Water mark).
}
\keyword{datasets}
