% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_ltla24}
\alias{boundaries_ltla24}
\title{Local Authority Districts (July 2024)}
\format{
A data frame of class "sf" with 361 rows and 3 variables:
\describe{
\item{ltla24_name}{LAD name}
\item{ltla24_code}{LAD code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_ltla24
}
\description{
A data set containing 2024 Local Authority Districts (LAD) geometries and
their associated names and codes. Boundaries are Ultra Generalised (500m) -
clipped to the coastline (Mean High Water mark).
}
\keyword{datasets}
