% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_msoa21_iz11_sdz21}
\alias{boundaries_msoa21_iz11_sdz21}
\title{Middle Layer Super Output Areas equivalents for the UK}
\format{
A data frame of class "sf" with 9,393 rows and 3 variables:
\describe{
\item{msoa21_name}{MSOA name}
\item{msoa21_code}{MSOA code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_msoa21_iz11_sdz21
}
\description{
A data set containing  Middle Super Output Area (MSOA) equivalent geometries and
their associated names and codes for the UK. Super generalised (200m) boundaries were
simplified using the Visvalingam algorithm set to a weighting of 0.7 with a
retaining proportion of 0.05 points.
}
\keyword{datasets}
