% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_pfa20}
\alias{boundaries_pfa20}
\title{Police Force Areas (2020)}
\format{
A data frame of class "sf" with 43 rows and 3 variables:
\describe{
\item{pfa20_name}{Police Force Area name}
\item{pfa20_code}{Police Force Area code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_pfa20
}
\description{
A data set containing 2020 Police Force Area geometries for England and Wales
and their associated names and codes.
}
\keyword{datasets}
