% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_region21}
\alias{boundaries_region21}
\title{Regions (December 2021)}
\format{
A data frame of class "sf" with 890 rows and 3 variables:
\describe{
\item{region21_name}{Region name}
\item{region21_code}{Region code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_region21
}
\description{
A data set containing December 2021 England Region geometries and
their associated names and codes.
}
\keyword{datasets}
