% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_sdz21}
\alias{boundaries_sdz21}
\title{Super Data Zones (2021)}
\format{
A data frame of class "sf" with 850 rows and 3 variables:
\describe{
\item{sdz21_name}{Super Data Zone name}
\item{sdz21_code}{Super Data Zone code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
boundaries_sdz21
}
\description{
A data set containing 2021 Northern Ireland Super Data Zones geometries and
their associated names and codes as used following the 2021 Census results.
}
\keyword{datasets}
