% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_soa11}
\alias{boundaries_soa11}
\title{Super Output Areas (2011)}
\format{
A data frame of class "sf" with 890 rows and 3 variables:
\describe{
\item{soa11_name}{Super Output Area name}
\item{soa11_code}{Super Output Area code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
boundaries_soa11
}
\description{
A data set containing 2011 Northern Ireland Super Output Area geometries and
their associated names and codes as used following the 2011 Census results.
}
\keyword{datasets}
