% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_stp21}
\alias{boundaries_stp21}
\title{Sustainability and Transformation Partnerships (April 2021)}
\format{
A data frame of class "sf" with 42 rows and 3 variables:
\describe{
\item{stp21_name}{STP name}
\item{stp21_code}{STP code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_stp21
}
\description{
A data set containing 2021 Sustainability and Transformation Partnerships
(STP) geometries and their associated names and codes. Boundaries are ultra
generalised (500m).
}
\keyword{datasets}
