% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_trusts_ni18}
\alias{boundaries_trusts_ni18}
\title{Northern Ireland Health and Social Care Trusts (2017)}
\format{
A data frame of class "sf" with 5 rows and 3 variables:
\describe{
\item{trust18_name}{Health and Social Care Trust name}
\item{trust18_code}{Health and Social Care Trust code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://www.opendatani.gov.uk/}
}
\usage{
boundaries_trusts_ni18
}
\description{
A data set containing 2017 Health and Social Care Trust geometries for
Northern Ireland and their associated names and codes.
}
\keyword{datasets}
