% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_vcsep22}
\alias{boundaries_vcsep22}
\title{VCSEP regions (2022).}
\format{
A data frame of class "tbl" with 5 rows and 2 variables:
\describe{
\item{vcsep_region}{VCSEP region name}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://github.com/VCS-Emergencies-Partnership/ad-hoc-analysis/tree/main/analysis/vcsep_regions_boundaries}
}
\usage{
boundaries_vcsep22
}
\description{
A data set containing geometries for the Voluntary and Community Sector
Emergency Partnership regions
}
\keyword{datasets}
