% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_wards21}
\alias{boundaries_wards21}
\title{Wards / Electoral Divisions (December 2021)}
\format{
A data frame of class "sf" with 8,694 rows and 3 variables:
\describe{
\item{ward21_name}{ward name}
\item{ward21_code}{ward code}
\item{geometry}{multipolygon coordinates}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
boundaries_wards21
}
\description{
A data set containing 2021 Ward / Electoral Division geometries and their
associated names and codes. Boundaries are Super generalised (200m) - clipped
to the coastline (Mean High Water mark).
}
\keyword{datasets}
