% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_extent}
\alias{calculate_extent}
\title{Calculate the 'extent' scores when aggregating up small areas}
\usage{
calculate_extent(
  data,
  var,
  higher_level_geography,
  population,
  weight_high_scores = TRUE
)
}
\arguments{
\item{data}{Data frame containing a variable to be aggregated, lower level
geography population estimates, and a higher level geographical
grouping variable}

\item{var}{Name of the variable in the data frame containing the variable to
be aggregated (e.g., score) for the lower level geography}

\item{higher_level_geography}{Name of the variable in the data frame
containing the higher level geography names/codes}

\item{population}{Name of the variable in the data frame containing
the population estimates of the lower level geography}

\item{weight_high_scores}{If TRUE higher scores are weighted, else lower
scores are weighted. For indicators like 'Alchol Misuse' and 'Ambulance Wait
Time' this should be set to TRUE. This is because higher values in these
outcomes indicate worse outcomes (higher vulnerability and lower capacity)
and this is where the weighting should be focused. For indicators like
'Physical Activity' and 'Bed Availability' it should be set to FALSE. This is
because lower values in these outcomes indicate worse outcomes (higher
vulnerability and lower capacity) and this is where the weighting should be
focused.}
}
\description{
"Extent" is the proportion of the local population that live in areas
classified as among the most deprived in the higher geography. The
calculation of extent is taken from the IMD technical report Appendix N:
}
\details{
"The population living in the most deprived 11 to 30 per cent of Lower-layer
Super Output Areas receive a sliding weight, ranging from 0.95 for those in
the most deprived eleventh percentile, to 0.05 for those in the most deprived
thirtieth percentile. In practice this means that the weight starts from 0.95
in the most deprived eleventh percentile, and then decreases by
(0.95-0.05)/19 for each of the subsequent nineteen percentiles until it
reaches 0.05 for the most deprived thirtieth percentile, and zero for areas
outside the most deprived 30 per cent"

The direction of the scale of data inputted by the functioin always matches
that of the data outputted. For example if data is inputted into the function
where high scores equals high vulnerability, the outputted data set will hold
this relationship true.
}
