% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate proportion of small areas in the higher-level geography that are
within the 10\% most deprived areas in the nation.}
\usage{
calculate_proportion(data, var, higher_level_geography)
}
\arguments{
\item{data}{Data frame containing a variable to be aggregated and a higher
level geographical grouping variable}

\item{var}{Name of the variable in the data frame for which you want to
calculate proportions. It must have only two possible values}

\item{higher_level_geography}{Name of the variable in the data frame
containing the higher level geography names/codes}
}
\description{
Calculate proportion of small areas in the higher-level geography that are
within the 10\% most deprived areas in the nation.
}
\examples{
\dontrun{
  calculate_proportion(IMD::imd_england_lsoa, IMD_decile, msoa_code, n_people)
}

}
