% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year.R
\name{filter_year}
\alias{filter_year}
\title{Filter data by year range}
\usage{
filter_year(data, start_year = 2000, end_year = 2020)
}
\arguments{
\item{data}{A data frame with a year column}

\item{start_year}{First year to include}

\item{end_year}{Last year to include}
}
\value{
A filtered data frame
}
\description{
Filter data by year range
}
