\name{find_flag}
\alias{find_flag}

\title{Identify national flags}

\description{Identify national flags}

\usage{
    find_flag(region = "Europa", lang = "en")
}

\arguments{
    \item{region}{A character (i.e. Asia). See \code{\link{get_region_name}} for the correct spelling. The default value is Europa.}
    \item{lang}{\emph{en} (english) or \emph{fr} (french)}
}

\value{A graphical device}

\author{Nicolas Casajus}

\seealso{
    \code{\link{get_region_name}},
    \code{\link{find_cities}},
    \code{\link{find_country}},
    \code{\link{learning}}
}

\examples{
### See https://github.com/ahasverus/geographr
}
