\name{geographr-package}

\alias{geographr-package}

\alias{geographr}

\docType{package}

\title{
    Discover the world geography with fun
}

\description{
    Discover the world geography with fun
}

\details{
    \tabular{ll}{
        Package: \tab geographr\cr
        Type: \tab Package\cr
        Version: \tab 1.0\cr
        Date: \tab 2016-06-02\cr
        License: \tab GPL (>= 3)\cr
    }
}

\author{
Nicolas Casajus

Maintainer: Nicolas Casajus <nicolas.casajus@gmail.com>
}
