\name{get_region_name}
\alias{get_region_name}

\title{Get the name of countries}

\description{Get the name of countries}

\usage{
    get_region_name(lang = "en")
}

\arguments{
    \item{lang}{\emph{en} (english) or \emph{fr} (french)}
}

\value{A vector with the right spelling of country names}

\author{Nicolas Casajus}

\seealso{
    \code{\link{get_country_name}},
    \code{\link{learning}},
    \code{\link{find_flag}},
    \code{\link{find_country}}
}

\examples{
### See https://github.com/ahasverus/geographr
}
