% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ccg21_stp21}
\alias{lookup_ccg21_stp21}
\title{Clinical Commissioning Group (2021) to Sustainability and Transformation
Partnership (2021) Lookup}
\format{
A data frame of class "tbl" with 106 rows and 6 variables:
\describe{
\item{ccg21_code}{CCG code}
\item{ccg21_nhs_code}{CCG's NHS code}
\item{ccg21_name}{CCG name}
\item{stp21_code}{STP code}
\item{stp21_nhs_code}{STP's NHS code}
\item{stp21_name}{STP name}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ccg21_stp21
}
\description{
A data set containing a lookup between CCG and STP in Scotland
}
\keyword{datasets}
