% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_dz01_dz11}
\alias{lookup_dz01_dz11}
\title{Data Zone (2001) to Data Zone (2011) Lookup}
\format{
A data frame of class "tbl" with 10,437 rows and 2 variables:
\describe{
\item{dz01_code}{Data Zone code (2001)}
\item{dz11_code}{Data Zone code (2011)}
...
}
}
\source{
\url{https://www.nrscotland.gov.uk}
}
\usage{
lookup_dz01_dz11
}
\description{
A data set containing a lookup between Data Zone 2001 and 2011 codes.
This data was constructed from the 2024-1 Scottish Postcode Directory.
}
\keyword{datasets}
