% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_dz11_iz11_ltla20}
\alias{lookup_dz11_iz11_ltla20}
\title{Data Zone (2011) to Intermediate Zone (2011) to LAD (December 2019) Lookup}
\format{
A data frame of class "tbl" with 6,976 rows and 6 variables:
\describe{
\item{dz11_name}{DZ name}
\item{dz11_code}{DZ code}
\item{iz11_name}{IZ name}
\item{iz11_code}{IZ code}
\item{ltla20_name}{LAD name}
\item{ltla20_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_dz11_iz11_ltla20
}
\description{
A data set containing a lookup between Data Zones, Intermediate Zones and Council Areas in Scotland
}
\keyword{datasets}
