% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_dz11_ltla19_hb19}
\alias{lookup_dz11_ltla19_hb19}
\title{Data Zone (2011) to Local Authority (2019) to Health Board (2019) Lookup}
\format{
A data frame of class "tbl" with 6,976 rows and 3 variables:
\describe{
\item{dz11_code}{Data zone (2011) code}
\item{ltla19_code}{Local authortiy (2019) code}
\item{hb19_code}{Health board (2019) code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_dz11_ltla19_hb19
}
\description{
A data set containing a lookup between Data Zones, Local Authorities (also
called Council Areas) and Health Boards in Scotland
}
\keyword{datasets}
