% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_dz21_sdz21_dea14_lgd14}
\alias{lookup_dz21_sdz21_dea14_lgd14}
\title{Data Zone (2021) to Super Data Zone (2021) to District Electoral Areas (2019)
to Local Government Districts (2014) Lookup}
\format{
A data frame of class "tbl" with 3,780 rows and 8 variables:
\describe{
\item{dz21_code}{Data zone (2021) code (comparable to Output Areas)}
\item{dz21_name}{Data zone (2021) name}
\item{sdz21_code}{Super data zone (2021) code (comparable to LSOA / Scottish Data Zone)}
\item{sdz21_name}{Super data zone (2021) name}
\item{dea14_code}{District Electoral Areas (2019) code}
\item{dea14_name}{District Electoral Areas (2019) name}
\item{lgd14_code}{Local Government District (2014) code (comparable to Local Authority)}
\item{lgd14_name}{Local Government District (2014) name}
...
}
}
\source{
\url{https://nisra.gov.uk/}
}
\usage{
lookup_dz21_sdz21_dea14_lgd14
}
\description{
A data set containing a lookup between Data Zones (Output Areas), Super Data Zones (LSOA),
District  Electoral Areas and Local Government Districts (Local Authority) in Northern Ireland
}
\keyword{datasets}
