% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lad_20_cauth_20}
\alias{lookup_lad_20_cauth_20}
\title{Local Authority Districts to Combined Authorities lookup (2020).}
\format{
A data frame of class "tbl" with 53 rows and 4 variables:
\describe{
\item{lad_20_name}{LAD name}
\item{lad_20_code}{LAD code}
\item{cauth_20_name}{County name}
\item{cauth_20_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lad_20_cauth_20
}
\description{
A data set containing a lookup between December 2020 Local Authority
Districts and Combined Authorities (CAUTH).
}
\keyword{datasets}
