% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lad_20_counties_ua_20}
\alias{lookup_lad_20_counties_ua_20}
\title{Local Authority Districts to County and Unitary Authorities lookup (2020).}
\format{
A data frame of class "tbl" with 336 rows and 4 variables:
\describe{
\item{lad_20_name}{LAD name}
\item{lad_20_code}{LAD code}
\item{county_ua_20_name}{County name}
\item{county_ua_20_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lad_20_counties_ua_20
}
\description{
A data set containing a lookup between December 2020 Local Authority
Districts and County and Unitary Authorities.
}
\keyword{datasets}
