% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lad_21_fra_21}
\alias{lookup_lad_21_fra_21}
\title{Local Authority Districts to Fire and Rescue Authorities lookup (2021).}
\format{
A data frame of class "tbl" with 331 rows and 4 variables:
\describe{
\item{lad_21_name}{LAD name}
\item{lad_21_code}{LAD code}
\item{fra_21_name}{FRA name}
\item{fra_21_code}{FRA code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lad_21_fra_21
}
\description{
A data set containing a lookup between April 2021 Local Authority
Fire and Rescue Authorities (FRA)
}
\keyword{datasets}
