% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lad_lad}
\alias{lookup_lad_lad}
\title{Changes to England Local Authority codes over time (from 2019).}
\format{
A data frame of class "tbl" with 382 rows and 6 variables:
\describe{
\item{lad_19_name}{2019 LAD name}
\item{lad_19_code}{2019 LAD code}
\item{lad_20_name}{2020 LAD name}
\item{lad_20_code}{2020 LAD code}
\item{lad_21_name}{2021 LAD name}
\item{lad_21_code}{2021 LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/datasets/code-history-database-december-2021-for-the-united-kingdom/about}
}
\usage{
lookup_lad_lad
}
\description{
A data set containing a lookup of the England Local Authority changes
between 2019, 2020 and 2021.
}
\keyword{datasets}
