% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa01_lsoa11}
\alias{lookup_lsoa01_lsoa11}
\title{Lower Layer Super Output Area 2001 to 2011 lookup}
\format{
A data frame of class "tbl" with 34,917 rows and 5 variables:
\describe{
\item{lsoa01_name}{LSOA 2001 name}
\item{lsoa01_code}{LSOA 2001 code}
\item{change_code}{U = no change; S = split; M = merged; X = an irregular
and fragmented relationship}
\item{lsoa11_name}{LSOA 2011 name}
\item{lsoa11_code}{LSOA 2011 code}
}
}
\usage{
lookup_lsoa01_lsoa11
}
\description{
A best fit lookup between lower layer super output areas as at December 2001
and lower layer super output areas and local authority districts as at December
2011 in England and Wales.
}
\keyword{datasets}
