% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa11_ccg21_stp21_ltla21}
\alias{lookup_lsoa11_ccg21_stp21_ltla21}
\title{Lower Layer Super Output Area (2011) to Clinical Commissioning Group (2021)
to Sustainability and Transformation Plan (2021) to Lower Tier Local
Authority District (2021) Lookup}
\format{
A data frame of class "tbl" with 32,844 rows and 8 variables:
\describe{
\item{lsoa11_name}{LSOA name}
\item{lsoa11_code}{LSOA code}
\item{ccg21_name}{CCG name}
\item{ccg21_code}{CCG code}
\item{stp21_name}{STP name}
\item{stp21_code}{STP code}
\item{ltla21_name}{LTLA name}
\item{ltla21_code}{LTLA code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa11_ccg21_stp21_ltla21
}
\description{
A data set containing a lookup between 2011 Lower Layer Super Output Areas to
2021 Clinical Commissioning Group areas to 2021 Sustainability and
Transformation Plan areas to 2021 Lower Tier Local Authority Districts.
}
\keyword{datasets}
