% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa11_lsoa21_ltla22}
\alias{lookup_lsoa11_lsoa21_ltla22}
\title{Lower Layer Super Output Area (2011) to Lower Layer Super Output Area (2021) to LAD (December 2022) Lookup}
\format{
A data frame of class "tbl" with 34,753 rows and 6 variables:
\describe{
\item{lsoa11_name}{LSOA 11 name}
\item{lsoa11_code}{LSOA 11 code}
\item{lsoa21_name}{LSOA 21 name}
\item{lsoa21_code}{LSOA 21 code}
\item{ltla22_name}{LAD name}
\item{ltla22_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa11_lsoa21_ltla22
}
\description{
A data set containing a lookup between 2011 Lower Super Output Areas (LSOA)
and 2021 Lower Super Output Areas (LSOA) and 2022 Local Authority Districts (LAD).
}
\keyword{datasets}
