% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa11_ltla20}
\alias{lookup_lsoa11_ltla20}
\title{Lower Layer Super Output Area (2011) to LAD (December 2020) Lookup}
\format{
A data frame of class "tbl" with 34,753 rows and 4 variables:
\describe{
\item{lsoa11_name}{LSOA name}
\item{lsoa11_code}{LSOA code}
\item{ltla20_name}{LAD name}
\item{ltla20_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa11_ltla20
}
\description{
A data set containing a lookup between 2011 Lower Super Output Areas (LSOA)
and 2020 Local Authority Districts (LAD).
}
\keyword{datasets}
