% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa11_ward17}
\alias{lookup_lsoa11_ward17}
\title{LSOA (2011) to Ward (2017) Lookup}
\format{
A data frame of class "tbl" with 34,753 rows and 2 variables:
\describe{
\item{lsoa11_code}{LSOA 2011 code}
\item{ward17_code}{Ward 2017 code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa11_ward17
}
\description{
A data set containing a lookup between 2011 LSOAs and 2021 wards in England
and Wales.
}
\keyword{datasets}
