% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa21_ward24_ltla24}
\alias{lookup_lsoa21_ward24_ltla24}
\title{Lower Layer Super Output Area (2021) to Ward (2024) to Local Authority District
(2024) Lookup}
\format{
A data frame of class "tbl" with 35,672 rows and 6 variables:
\describe{
\item{lsoa21_name}{LSOA name}
\item{lsoa21_code}{LSOA code}
\item{ward24_name}{Ward name}
\item{ward24_code}{Ward code}
\item{ltla24_name}{LAD name}
\item{ltla24_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa21_ward24_ltla24
}
\description{
A data set containing a lookup between 2021 Lower Super Output Areas (LSOA)
and 2024 Wards and Local Authority Districts.
}
\keyword{datasets}
