% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_lsoa_lad_21}
\alias{lookup_lsoa_lad_21}
\title{Lower Layer Super Output Area (2011) to LAD (December 2021) Lookup}
\format{
A data frame of class "tbl" with 34,753 rows and 4 variables:
\describe{
\item{lsoa_11_name}{LSOA name}
\item{lsoa_11_code}{LSOA code}
\item{lad_21_name}{LAD name}
\item{lad_21_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_lsoa_lad_21
}
\description{
A data set containing a lookup between 2011 Lower Super Output Areas (LSOA)
and 2021 Local Authority Districts (LAD).
}
\keyword{datasets}
