% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla19_cauth19}
\alias{lookup_ltla19_cauth19}
\title{Local Authority Districts to Combined Authorities lookup (2019).}
\format{
A data frame of class "tbl" with 53 rows and 4 variables:
\describe{
\item{ltla19_name}{LAD name}
\item{ltla19_code}{LAD code}
\item{cauth19_name}{County name}
\item{cauth19_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla19_cauth19
}
\description{
A data set containing a lookup between December 2019 Local Authority
Districts and Combined Authorities (CAUTH).
}
\keyword{datasets}
