% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla19_region19}
\alias{lookup_ltla19_region19}
\title{Local Authority District to Region (December 2019) Lookup in England}
\format{
A data frame of class "tbl" with 317 rows and 4 variables:
\describe{
\item{ltla19_name}{LAD name}
\item{ltla19_code}{LAD code}
\item{region19_name}{Region name}
\item{region19_code}{Region code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla19_region19
}
\description{
A data set containing a lookup between 2019 Local Authority Districts (LAD)
and 2019 English Regions.
}
\keyword{datasets}
