% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla19_utla19}
\alias{lookup_ltla19_utla19}
\title{Local Authority Districts to County and Unitary Authorities lookup (2019).}
\format{
A data frame of class "tbl" with 339 rows and 4 variables:
\describe{
\item{ltla19_name}{LAD name}
\item{ltla19_code}{LAD code}
\item{utla19_name}{County name}
\item{utla19_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla19_utla19
}
\description{
A data set containing a lookup between December 2019 Local Authority
Districts and County and Unitary Authorities.
}
\keyword{datasets}
