% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla20_region20}
\alias{lookup_ltla20_region20}
\title{Local Authority District to Region (December 2020) Lookup in England}
\format{
A data frame of class "tbl" with 314 rows and 4 variables:
\describe{
\item{ltla20_name}{LAD name}
\item{ltla20_code}{LAD code}
\item{region20_name}{Region name}
\item{region20_code}{Region code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla20_region20
}
\description{
A data set containing a lookup between 2020 Local Authority Districts (LAD)
and 2020 English Regions.
}
\keyword{datasets}
