% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_brc}
\alias{lookup_ltla21_brc}
\title{Local Authority District (2021) to British Red Cross area lookup}
\format{
A data frame of class "tbl" with 382 rows and 2 variables:
\describe{
\item{ltla21_code}{LAD code}
\item{brc_area}{British Red Cross area name}
...
}
}
\source{
\url{https://redcross.org.uk/}
}
\usage{
lookup_ltla21_brc
}
\description{
A data set containing a lookup between UK Local Authority Districts and
British Red Cross operational areas.
}
\keyword{datasets}
