% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_cauth21}
\alias{lookup_ltla21_cauth21}
\title{Local Authority Districts to Combined Authorities lookup (2021).}
\format{
A data frame of class "tbl" with 53 rows and 4 variables:
\describe{
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
\item{cauth21_name}{County name}
\item{cauth21_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla21_cauth21
}
\description{
A data set containing a lookup between December 2021 Local Authority
Districts and Combined Authorities (CAUTH).
}
\keyword{datasets}
