% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_hsct18}
\alias{lookup_ltla21_hsct18}
\title{Local Government District (2021) to Health and Social Care Trust (2018) Lookup}
\format{
A data frame of class "tbl" with 11 rows and 4 variables:
\describe{
\item{ltla21_name}{Local Government District (2021) name}
\item{ltla21_code}{Local Government District (2021) code}
\item{trust18_name}{HHealth and Social Care Trust (2018) name}
\item{trust18_code}{HHealth and Social Care Trust (2018) code}
...
}
}
\source{
\url{https://www.nisra.gov.uk/sites/nisra.gov.uk/files/publications/Maps-of-NI-Health-Trusts-and-Local-Government-Districts.pdf}
}
\usage{
lookup_ltla21_hsct18
}
\description{
A data set containing a lookup between Local Authorities (also
called Local Government Districts) and Health and Social Care Trusts in Northern Ireland
}
\keyword{datasets}
