% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_lhb22}
\alias{lookup_ltla21_lhb22}
\title{Wales Local Authority Districts (2021) to Local Health Boards (2022) Lookup}
\format{
A data frame of class "tbl" with 22 rows and 4 variables:
\describe{
\item{ltla21_code}{Local Government District (2021) code}
\item{ltla21_name}{Local Government District (2021) name}
\item{lhb22_code}{Local Health Board (2022) code}
\item{lhb22_name}{Local Health Board (2022) name}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla21_lhb22
}
\description{
A data set containing a lookup between Local Authorities (2021) and Local
Health Boards (December 2022) in Wales.
}
\keyword{datasets}
