% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_region21}
\alias{lookup_ltla21_region21}
\title{Local Authority District to Region (April 2021) Lookup in England}
\format{
A data frame of class "tbl" with 317 rows and 4 variables:
\describe{
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
\item{region21_name}{Region name}
\item{region21_code}{Region code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla21_region21
}
\description{
A data set containing a lookup between 2021 Local Authority Districts (LAD)
and 2021 English Regions.
}
\keyword{datasets}
