% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_utla21}
\alias{lookup_ltla21_utla21}
\title{Local Authority Districts to County and Unitary Authorities lookup (2021).}
\format{
A data frame of class "tbl" with 331 rows and 4 variables:
\describe{
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
\item{utla21_name}{County name}
\item{utla21_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla21_utla21
}
\description{
A data set containing a lookup between December 2021 Local Authority
Districts and County and Unitary Authorities.
}
\keyword{datasets}
