% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla21_vcsep22}
\alias{lookup_ltla21_vcsep22}
\title{Local Authority Districts to VCSEP regions lookup (2021).}
\format{
A data frame of class "tbl" with 309 rows and 3 variables:
\describe{
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
\item{vcsep_region}{VCSEP region name}
...
}
}
\source{
\url{https://github.com/VCS-Emergencies-Partnership/ad-hoc-analysis/tree/main/analysis/vcsep_regions_boundaries}
}
\usage{
lookup_ltla21_vcsep22
}
\description{
A data set containing a lookup between December 2021 Local Authority
Districts and Voluntary and Community Sector Emergency Partnership regions
}
\keyword{datasets}
