% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla22_ltla23}
\alias{lookup_ltla22_ltla23}
\title{Local Authority Districts (2022) to Local Authority Districts (2023).}
\format{
A data frame of class "tbl" with 374 rows and 4 variables:
\describe{
\item{ltla22_code}{LAD 22 code}
\item{ltla22_name}{LAD 22 name}
\item{ltla23_code}{LAD 23 code}
\item{ltla23_name}{LAD 23 name}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla22_ltla23
}
\description{
A data set containing a lookup between 2022 Local Authority
Districts and 2023 Local Authority Districts.
}
\keyword{datasets}
