% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla23_cauth23}
\alias{lookup_ltla23_cauth23}
\title{Local Authority Districts to Combined Authorities lookup (2023).}
\format{
A data frame of class "tbl" with 53 rows and 4 variables:
\describe{
\item{ltla23_name}{LAD name}
\item{ltla23_code}{LAD code}
\item{cauth23_name}{County name}
\item{cauth23_code}{County code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla23_cauth23
}
\description{
A data set containing a lookup between December 2023 Local Authority
Districts and Combined Authorities (CAUTH).
}
\keyword{datasets}
