% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla23_region23}
\alias{lookup_ltla23_region23}
\title{Local Authority District to Region (April 2023) Lookup in England}
\format{
A data frame of class "tbl" with 296 rows and 4 variables:
\describe{
\item{ltla23_name}{LAD name}
\item{ltla23_code}{LAD code}
\item{region23_name}{Region name}
\item{region23_code}{Region code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla23_region23
}
\description{
A data set containing a lookup between 2023 Local Authority Districts (LAD)
and 2023 English Regions.
}
\keyword{datasets}
