% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla23_utla23}
\alias{lookup_ltla23_utla23}
\title{Lower Tier Local Authority (2023) to Upper Tier Local Authority (2023).}
\format{
A data frame of class "tbl" with 318 rows and 4 variables:
\describe{
\item{ltla23_code}{LTLA code}
\item{ltla23_name}{LTLA name}
\item{utla23_code}{UTLA code}
\item{utla23_name}{UTLA name}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla23_utla23
}
\description{
A data set containing a lookup between Lower Tier Local Authority (2023) to
Upper Tier Local Authority (2023).
}
\keyword{datasets}
