% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla24_fra24}
\alias{lookup_ltla24_fra24}
\title{Local Authority Districts to Fire and Rescue Authorities lookup (2024).}
\format{
A data frame of class "tbl" with 318 rows and 4 variables:
\describe{
\item{ltla24_name}{LAD name}
\item{ltla24_code}{LAD code}
\item{fra24_name}{FRA name}
\item{fra24_code}{FRA code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla24_fra24
}
\description{
A data set containing a lookup between April 2024 Local Authority
Fire and Rescue Authorities (FRA)
}
\keyword{datasets}
