% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ltla24_region24}
\alias{lookup_ltla24_region24}
\title{Local Authority District to Region (April 2024) Lookup in England}
\format{
A data frame of class "tbl" with 296 rows and 4 variables:
\describe{
\item{ltla24_name}{LAD name}
\item{ltla24_code}{LAD code}
\item{region24_name}{Region name}
\item{region24_code}{Region code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ltla24_region24
}
\description{
A data set containing a lookup between 2024 Local Authority Districts (LAD)
and 2024 English Regions.
}
\keyword{datasets}
