% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_msoa11_ltla21}
\alias{lookup_msoa11_ltla21}
\title{Middle Layer Super Output Area (2011) to LAD (December 2021) Lookup}
\format{
A data frame of class "tbl" with 7,201 rows and 4 variables:
\describe{
\item{msoa11_name}{MSOA name}
\item{msoa11_code}{MSOA code}
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_msoa11_ltla21
}
\description{
A data set containing a lookup between 2011 Middle Super Output Areas (MSOA)
and 2021 Local Authority Districts (LAD).
}
\keyword{datasets}
