% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_msoa11_msoa21_ltla22}
\alias{lookup_msoa11_msoa21_ltla22}
\title{Middle Layer Super Output Area (2011) to Middle Layer Super Output Area (2021) to LAD (December 2022) Lookup}
\format{
A data frame of class "tbl" with 7.201 rows and 6 variables:
\describe{
\item{msoa11_name}{MSOA 11 name}
\item{msoa11_code}{MSOA 11 code}
\item{msoa21_name}{MSOA 21 name}
\item{msoa21_code}{MSOA 21 code}
\item{ltla22_name}{LAD name}
\item{ltla22_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_msoa11_msoa21_ltla22
}
\description{
A data set containing a lookup between 2011 Middle Super Output Areas (MSOA)
and 2021 Middle Super Output Areas (MSOA) and 2022 Local Authority Districts (LAD).
}
\keyword{datasets}
