% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_msoa11_ward17}
\alias{lookup_msoa11_ward17}
\title{MSOA (2011) to Ward (2017) Lookup}
\format{
A data frame of class "tbl" with 13,873 rows and 2 variables:
\describe{
\item{msoa11_code}{MSOA 2011 code}
\item{ward17_code}{Ward 2017 code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_msoa11_ward17
}
\description{
A data set containing a lookup between 2011 MSOAs and 2021 wards in England
and Wales.
}
\keyword{datasets}
