% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_msoa_lad_20}
\alias{lookup_msoa_lad_20}
\title{Middle Layer Super Output Area (2011) to LAD (December 2020) Lookup}
\format{
A data frame of class "tbl" with 7,201 rows and 4 variables:
\describe{
\item{msoa_11_name}{MSOA name}
\item{msoa_11_code}{MSOA code}
\item{lad_20_name}{LAD name}
\item{lad_20_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_msoa_lad_20
}
\description{
A data set containing a lookup between 2011 Middle Super Output Areas (MSOA)
and 2020 Local Authority Districts (LAD).
}
\keyword{datasets}
