% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_nhs_trusts22_icb22}
\alias{lookup_nhs_trusts22_icb22}
\title{England NHS Trusts (2022) to ICB (2022) Lookup}
\format{
A data frame of class "tbl" with 1,321 rows and 3 variables:
\describe{
\item{icb22_code}{ICB code}
\item{nhs_trust22_code}{NHS Trust code}
\item{proportion_trust_came_from_icb}{The proportion of patients from all admissions from a given Trust that came from a given ICB}
...
}
}
\source{
\url{https://app.box.com/s/qh8gzpzeo1firv1ezfxx2e6c4tgtrudl}
}
\usage{
lookup_nhs_trusts22_icb22
}
\description{
A data set containing a lookup between 2022 England NHS Trusts and 2022
Integrate Care Boards (ICBs).
}
\section{Copyright reference}{
Office for Health Improvement & Disparities. NHS Acute (Hospital) Trust
Catchment Populations. 18.10.2022. https://AcuteTrustCatchmentPopulation/app.powerbi.com
© Crown copyright 2022
}

\keyword{datasets}
