% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_nhs_trusts22_ltla21}
\alias{lookup_nhs_trusts22_ltla21}
\title{England NHS Trusts (2022) to LTLA (2021) Lookup}
\format{
A data frame of class "tbl" with 574,104 rows and 4 variables:
\describe{
\item{ltla21_code}{LTLA code}
\item{nhs_trust22_code}{NHS Trust code}
\item{proportion_trust_came_from_ltla}{The proportion of patients from all admissions from a given Trust that came from a given LTLA}
...
}
}
\source{
\url{https://app.box.com/s/qh8gzpzeo1firv1ezfxx2e6c4tgtrudl}
}
\usage{
lookup_nhs_trusts22_ltla21
}
\description{
A data set containing a lookup between 2022 England NHS Trusts and 2021 LTLA
areas.
}
\section{Copyright reference}{
Office for Health Improvement & Disparities. NHS Acute (Hospital) Trust
Catchment Populations. 18.10.2022. https://AcuteTrustCatchmentPopulation/app.powerbi.com
© Crown copyright 2022
}

\keyword{datasets}
