% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_nhs_trusts22_nhs_region21}
\alias{lookup_nhs_trusts22_nhs_region21}
\title{NHS Trusts (February 2022) to NHS Regions (April 2021) Lookup}
\format{
A data frame of class "tbl" with 214 rows and 6 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust code}
\item{nhs_trust22_name}{NHS Trust name}
\item{nhs_region21_nhs_code}{NHS Region code (NHS version)}
\item{nhs_region21_ons_code}{NHS Region code (ONS version)}
\item{nhs_region21_name}{NHS Region name}
\item{status}{whether the Trust is 'open' or 'closed'}
...
}
}
\source{
\url{https://digital.nhs.uk/}
}
\usage{
lookup_nhs_trusts22_nhs_region21
}
\description{
A data set containing February 2022 NHS Trusts and their associated names,
codes, and open status, with a lookup to NHS Regions (April 2021).
}
\keyword{datasets}
