% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_nhs_trusts22_stp21}
\alias{lookup_nhs_trusts22_stp21}
\title{England NHS Trusts (2022) to STP/ICS (2021) Lookup}
\format{
A data frame of class "tbl" with 209 rows and 4 variables:
\describe{
\item{nhs_trust22_name}{NHS Trust name}
\item{nhs_trust22_code}{NHS Trust code}
\item{stp21_name}{STP name}
\item{stp21_code}{STP code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_nhs_trusts22_stp21
}
\description{
A data set containing a lookup between 2022 England NHS Trusts and 2021 STP
areas.
}
\keyword{datasets}
