% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_postcode_dz21_sdz21_dea14_lgd14}
\alias{lookup_postcode_dz21_sdz21_dea14_lgd14}
\title{Postcode to Data Zone to Super DataZone to District Electoral Areas to Local Government District
lookup for Northern Ireland}
\format{
A data frame of class "tbl":
\describe{
\item{postcode}{Postcode}
\item{dz21_code}{Data zone (2021) code (comparable to Output Areas)}
\item{dz21_name}{Data zone (2021) name}
\item{sdz21_code}{Super data zone (2021) code (comparable to LSOA / Scottish Data Zone)}
\item{sdz21_name}{Super data zone (2021) name}
\item{dea14_code}{District Electoral Areas (2019) code}
\item{dea14_name}{District Electoral Areas (2019) name}
\item{lgd14_code}{Local Government District (2014) code (comparable to Local Authority)}
\item{lgd14_name}{Local Government District (2014) name}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_postcode_dz21_sdz21_dea14_lgd14
}
\description{
Postcode to Data Zone to Super DataZone to District Electoral Areas to Local Government District
lookup for Northern Ireland
}
\keyword{datasets}
