% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_postcode_oa21_lsoa21_msoa21_ltla22}
\alias{lookup_postcode_oa21_lsoa21_msoa21_ltla22}
\title{Postcode to Output Area (2021) to Lower Layer Super Output Area (2021) to
Middle Layer Super Output Area (2021) to Local Authority District (2022)
Lookup}
\format{
A data frame of class "tbl" with 2,600,236 rows and 5 variables:
\describe{
\item{postcode}{Postcodes}
\item{oa21_code}{OA code}
\item{lsoa21_code}{LSOA code}
\item{msoa21_code}{MSOA code}
\item{ltla22_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_postcode_oa21_lsoa21_msoa21_ltla22
}
\description{
A data set containing a lookup between postcodes, OAs, LSOAs, MSOAs, and
2022 Local Authority Districts, for England, Wales, and Scotland.
}
\keyword{datasets}
