% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_postcode_oa_lsoa_msoa_lad_20}
\alias{lookup_postcode_oa_lsoa_msoa_lad_20}
\title{Postcode to Output Area (2011) to Lower Layer Super Output Area (2011) to
Middle Layer Super Output Area (2011) to Local Authority District (2020)
Lookup}
\format{
A data frame of class "tbl" with 2,568,780 rows and 5 variables:
\describe{
\item{postcode}{Postcodes}
\item{oa_11_code}{OA code}
\item{lsoa_11_code}{LSOA code}
\item{msoa_11_code}{MSOA code}
\item{lad_20_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_postcode_oa_lsoa_msoa_lad_20
}
\description{
A data set containing a lookup between postcodes, OAs, LSOAs, MSOAs, and
2020 Local Authority Districts, for England, Wales, and Scotland.
}
\keyword{datasets}
