% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_sa11_soa11_lgd18}
\alias{lookup_sa11_soa11_lgd18}
\title{Small Areas (2011) to SOAs to Local Government Districts (December 2018)
Lookup with Area Classifications in Northern Ireland}
\format{
A data frame of class "tbl" with 4,537 rows and 5 variables:
\describe{
\item{sa11_code}{Small Area code}
\item{soa11_name}{Super Output Area name}
\item{soa11_code}{Super Output Area code}
\item{lgd18_name}{Local Government District name}
\item{lgd18_code}{Local Government District code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_sa11_soa11_lgd18
}
\description{
A data set containing a lookup between Small Areas, Super Output Areas,
and Local Government Districts in Northern Ireland
}
\keyword{datasets}
