% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_soa01_lgd14}
\alias{lookup_soa01_lgd14}
\title{SOA (2001) to LGD (2014) Lookup}
\format{
A data frame of class "tbl" with 890 rows and 3 variables:
\describe{
\item{soa01_code}{Super Output Area 2001 code}
\item{lgd14_code}{Local Government District 2014 code}
\item{lgd14_name}{Local Government District 2014 name}
...
}
}
\source{
\url{https://www.nisra.gov.uk/publications/geography-lookup-tables}
}
\usage{
lookup_soa01_lgd14
}
\description{
A data set containing a lookup between SOA and LGD in Northern Ireland.
}
\keyword{datasets}
