% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_trust_trust}
\alias{lookup_trust_trust}
\title{NHS Trust and Orginisation Changes}
\format{
A data frame of class "tbl" with 25,480 rows and 3 variables:
\describe{
\item{old_code}{The old organisation code}
\item{new_code}{The new organisation code}
\item{date}{The date of the change}
...
}
}
\source{
\url{https://digital.nhs.uk/}
}
\usage{
lookup_trust_trust
}
\description{
A data set containing a lookup of NHS Trust and organisational changes
}
\keyword{datasets}
