% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_ward21_ltla21}
\alias{lookup_ward21_ltla21}
\title{Ward (December 2021) to LTLA (December 2021) Lookup}
\format{
A data frame of class "tbl" with 8,694 rows and 4 variables:
\describe{
\item{ward21_name}{MSOA name}
\item{ward21_code}{MSOA code}
\item{ltla21_name}{LAD name}
\item{ltla21_code}{LAD code}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
lookup_ward21_ltla21
}
\description{
A data set containing a lookup between 2021 wards and 2021 Lower Tier Local
Authority Districts (LAD).
}
\keyword{datasets}
