% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_codes_hb19}
\alias{names_codes_hb19}
\title{Names and codes for NHS Scotland Health Boards (2014 - 2019)}
\format{
A data frame of class "tbl" with 18 rows and 4 variables:
\describe{
\item{hb19_name}{Health Board name}
\item{hb19_code}{Health Board code}
\item{date_enacted}{Date Health Board Code was enacted}
\item{date_archived}{Date Health Board Code was archived}
...
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
names_codes_hb19
}
\description{
Names and codes for NHS Scotland Health Boards (2014 - 2019)
}
\keyword{datasets}
