% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tenure_ridge.R
\name{plot_tenure_ridge}
\alias{plot_tenure_ridge}
\title{Plot presidential tenure by selected regime types}
\usage{
plot_tenure_ridge(regimes, data = load_data())
}
\arguments{
\item{regimes}{A character vector of regime categories to include}

\item{data}{A data frame containing presidential tenure and regime data}
}
\value{
A ggplot object.
}
\description{
Creates a density ridge plot showing the distribution of presidential tenure
for specific regime types
}
