% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_hospitals22}
\alias{points_hospitals22}
\title{NHS England Hopsitals (July 2022)}
\format{
A data frame of class "sf" with 1,061 rows and 2 variables:
\describe{
\item{hospital22_name}{NHS England hospital}
\item{geometry}{point coordinates}
...
}
}
\source{
\url{https://digital.nhs.uk/}
}
\usage{
points_hospitals22
}
\description{
A data set containing July 2022 NHS England hospital point geometries (i.e.,
lat/long coordinates) and their associated names.
}
\keyword{datasets}
