% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_nhs_trusts22}
\alias{points_nhs_trusts22}
\title{NHS Trusts (February 2022)}
\format{
A data frame of class "sf" with 218 rows and 4 variables:
\describe{
\item{nhs_trust22_name}{NHS Trust name}
\item{nhs_trust22_code}{NHS Trust code}
\item{status}{whether the Trust is 'open' or 'closed'}
\item{geometry}{point coordinates}
...
}
}
\source{
\url{https://digital.nhs.uk/}
}
\usage{
points_nhs_trusts22
}
\description{
A data set containing February 2022 NHS Trusts point geometries (i.e.,
lat/long coordinates) and their associated names, codes, and open status.
}
\keyword{datasets}
