% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_scotland_hospitals24}
\alias{points_scotland_hospitals24}
\title{NHS Scotland Hopsitals (2024)}
\format{
A data frame of class "sf" with 309 rows and 7 variables:
\describe{
\item{scotland_hospital24_code}{NHS Scotland hospital code}
\item{scotland_hospital24_name}{NHS Scotland hospital name}
\item{hb19_code}{Health Board code}
\item{hb19_name}{Health Board name}
\item{postcode}{Postcode}
\item{address}{Address}
\item{geometry}{point coordinates}
...
}
}
\source{
\url{https://data.spatialhub.scot/}
}
\usage{
points_scotland_hospitals24
}
\description{
A data set containing 2024 NHS Scotland hospital point geometries (i.e.,
lat/long coordinates) and their associated names and addresses.
}
\keyword{datasets}
