% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc11_msoa11}
\alias{ruc11_msoa11}
\title{Rural-Urban Classifications for Middle Layer Super Output Areas (2011) in
England and Wales}
\format{
A data frame of class "tbl" with 7,201 rows and 4 variables:
\describe{
\item{msoa11_code}{MSOA code}
\item{ruc11_code}{Rural-Urban Classification code}
\item{ruc}{Whether the MSOA is classified as 'Rural' or 'Urban'}
\item{classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
ruc11_msoa11
}
\description{
A data set containing rural-urban classifications for MSOAs in England and
Wales.
}
\keyword{datasets}
