% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc21_lsoa21}
\alias{ruc21_lsoa21}
\title{Rural-Urban Classifications for Lower Layer Super Output Areas (2021) in
England and Wales}
\format{
A data frame of class "tbl" with 35,672 rows and 4 variables:
\describe{
\item{lsoa21_code}{LSOA code}
\item{ruc21_code}{Rural-Urban Classification code}
\item{ruc}{Whether the LSOA is classified as 'Rural' or 'Urban'}
\item{classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
ruc21_lsoa21
}
\description{
A data set containing rural-urban classifications for LSOAs in England and
Wales.
}
\keyword{datasets}
