% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc21_ltla24}
\alias{ruc21_ltla24}
\title{Rural-Urban Classifications for Local Authorities (2024) in England and Wales}
\format{
A data frame of class "tbl" with 318 rows and 4 variables:
\describe{
\item{ltla24_code}{Local Authority code}
\item{ruc21_code}{Rural-Urban Classification code}
\item{ruc}{Whether the Local Authority is classified as 'Rural' or 'Urban'}
\item{classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
ruc21_ltla24
}
\description{
A data set containing rural-urban classifications for Local Authorities in England and
Wales.
}
\keyword{datasets}
