% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc21_msoa21}
\alias{ruc21_msoa21}
\title{Rural-Urban Classifications for Middle Layer Super Output Areas (2021) in
England and Wales}
\format{
A data frame of class "tbl" with 7,264 rows and 4 variables:
\describe{
\item{msoa21_code}{MSOA code}
\item{ruc21_code}{Rural-Urban Classification code}
\item{ruc}{Whether the MSOA is classified as 'Rural' or 'Urban'}
\item{classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
ruc21_msoa21
}
\description{
A data set containing rural-urban classifications for MSOAs in England and
Wales.
}
\keyword{datasets}
