% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc_dz_11}
\alias{ruc_dz_11}
\title{Rural-Urban Classifications for Data Zones (2011) in Scotland}
\format{
A data frame of class "tbl" with 6,976 rows and 4 variables:
\describe{
\item{dz_11_code}{Data Zone code}
\item{ruc_16_code}{Rural-Urban Classification code}
\item{RUC}{Whether the DZ is classified as 'Rural' or 'Urban'}
\item{Classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://www.gov.scot/}
}
\usage{
ruc_dz_11
}
\description{
A data set containing rural-urban classifications for Data Zones in Scotland.
}
\keyword{datasets}
