% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruc_lsoa_11}
\alias{ruc_lsoa_11}
\title{Rural-Urban Classifications for Lower Layer Super Output Areas (2011) in
England and Wales}
\format{
A data frame of class "tbl" with 34,753 rows and 4 variables:
\describe{
\item{lsoa_11_code}{LSOA code}
\item{ruc_11_code}{Rural-Urban Classification code}
\item{RUC}{Whether the LSOA is classified as 'Rural' or 'Urban'}
\item{Classification}{A more detailed description of the classification}
...
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
ruc_lsoa_11
}
\description{
A data set containing rural-urban classifications for LSOAs in England and
Wales.
}
\keyword{datasets}
