% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run_api}
\alias{run_api}
\title{Run API}
\usage{
run_api(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run API
}
\examples{
\dontrun{
run_api()
run_api(swagger = TRUE, port = 8000)
}
}
