% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_lsoa.R
\name{aggregate_lba_lsoas}
\alias{aggregate_lba_lsoas}
\title{Aggregate LSOAs'11 - LSOA'21 for Left Behind Areas}
\usage{
aggregate_lba_lsoas(data)
}
\arguments{
\item{data}{Dataframe containing 2011 LSOA codes to be aggregated to 2021 LSOA codes}
}
\value{
Dataframe containing the aggregated 2021 LSOA codes
}
\description{
Changes were made to to 2011 LSOAs in 2021 to reflect population and household
changes. Change codes:
\itemize{
\item U = Unchanged - 31,810 LSOAs remained unchanged from 2011 to 2021
\item S = Split - 834 LSOAs were split from 2011 to 2021 into two or more
\item M = Merged - 194 LSOAs were merged from 2011 to 2021 with one or more
\item X = Fragmented - 6 LSOAs were fragmented/redesigned from 2011 to 2021
}
}
\details{
Aggregation strategy going from 2021 codes to 2011 codes:
\itemize{
\item change_code == "U": no action required
\item change_code == "S": if ANY of the 2021 LSOAs resulting from the split is
left behind, 2011 LSOA is considered left-behind
\item change_code == "M": 2011 LSOA's inherit the left-behind characteristics of
the 2021 LSOA that resulted from the merge: if the 2021 LSOA is left behind,
all corresponding 2011 LSOA's are considered left-behind, and vice versa
\item change_code == "X": 2011 LSOA's inherit the left behind characteristics of
their corresponding 2021 LSOA. Then group by 2011 LSOA, if any is
left-behind, consider the LSOA left-behind
}
}
