% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load democracy data}
\usage{
load_data()
}
\value{
A tibble containing country-level democracy data.
}
\description{
Loads the TidyTuesday democracy dataset from November 5, 2024.
}
